[GGE::TreeSpider] nested repeated captures, take I
We now handle `( (.) )*` correctly. The outer parens create matches which
end up in an Array. The inner parens create matches all numbered 0 inside
those outer matches.

We also handle `( [.] )*` correctly. But not `[ (.) ]*`. That case still
confuses me. Seems it should work just like `( [.] )*`. I need to consult
S05 as to why and PGE as to how.