[GGE::TreeSpider] nested repeated captures, take II
This commit fixes test 24 in rx_captures (`( (.) )*`), but makes test 23
(`( [.] )*`), fixed by the previous commit, fail. I've a feeling that this
commit and the previous one both apply a fix to the exact same use case,
resulting in double arrays or something. Will try and reverse the effects
of the previous commits to see if that helps.