[GGE::Perl6Regex] introduced no-scope CGroups
This helps emulate Perl 5 numbering, as in /$1=[(.) (.)] (.)/. This creates
four captures $1, $2, $3, and $4. $1 corresponds to [(.) (.)]. It's a
Group syntactically, but it gets forced into being a CGroup, and has its
'isscope' property set to False. Thus, the $2 and $3 captures, despite
being nested directly inside this CGroup, end up being attached to the
level above, the entire match object.