[GGE::Perl6Regex] implemented &before
This was possibly the hardest single commit so far, in terms of the time
it took and the thinking it required. Especially since it doesn't do very
much. A 'before' subrule call anchors the match at that point, checks if
it matches, and returns either a failed match or a zero-width successful
match.