[GGE::Perl6Regex] check .Bool rather than .defined
The code was the way it was because it turns out that Nil is defined in
Rakudo, whereas it should be undefined. The solution employed involved
descending into some nodes twice, which caused other problems. The solution
new employed simply checks for boolean truth... which should work in all
cases. I think.