[GGE::Perl6Regex] testing definedness
There's been a problem with testing values for definedness, when the undefined
value being returned is Nil, because Nil when stored in a variable mutates
into a defined value. Instead, we return Mu for the time being, a solution
which is impure but workable.

GGE is now fully ported to Rakudo master! \o/